+++ src/models.rs
@@ -4,7 +4,7 @@ use crate::*;
 
 #[derive(Default, Debug, Serialize, Deserialize)]
 pub struct Config {
-    pub translations: Vec<u8>,
+    pub translations: Vec<u16>,
     pub arabic : bool,
 }
